fix: invoke metadata endpoint on action configu...
...ration finish

This adds invocation of metadata endpoint when user is done defining
action properties in order to make sure that the input/output data
shapes are set with all the properties defined.

No data shape would be set for action whose properties are defined in
single step otherwise.

Needs error handling.